+++ src/lwc/geTemplateBuilder/geTemplateBuilder.html
@@ -57,8 +57,10 @@
                                                            form-sections={formSections}
                                                            active-form-section-id={activeFormSectionId}
                                                            previous-save-attempted={previousSaveAttempted}
+                                                           section-ids-by-field-mapping-developer-names={sectionIdsByFieldMappingDeveloperNames}
                                                            onupdatevalidity={handleUpdateValidity}
                                                            ontogglemodal={toggleModal}
+                                                           ontogglefieldmapping={handleToggleFieldMapping}
                                                            onchangeactivesection={handleChangeActiveSection}
                                                            onaddformsection={handleAddFormSection}
                                                            onformsectionup={handleFormSectionUp}
